Changed Timeout attributes to int64_t type

+++ src/Controller.h
@@ -12,10 +12,10 @@
 class Controller {
     private:
         // System Timeouts 
-        float DOOR_UNLOCK_TIMEOUT = 10000; // Random value for now
-        float ALARM_TIMEOUT = 10000; // Random value for now
-        float PIN_ENTRY_TIMEOUT = 10000; // Random value for now
-        float MOTION_SENSOR_TIMEOUT = 10000; // Random value for now
+        int64_t DOOR_UNLOCK_TIMEOUT = 10000; // Random value for now
+        int64_t ALARM_TIMEOUT = 10000; // Random value for now
+        int64_t PIN_ENTRY_TIMEOUT = 10000; // Random value for now
+        int64_t MOTION_SENSOR_TIMEOUT = 10000; // Random value for now
 
         // Sensors
         MagneticSensor *door;
